last insert id is long, not int. Fix that in tntdb::Connection
git-svn-id: svn://svn.code.sf.net/p/tntdb/code/trunk@178 8aebba32-5d12-0410-a889-bbfed4e15866

+++ tntdb/include/tntdb/bits/connection.h
@@ -164,7 +164,7 @@ namespace tntdb
       /**
        * returns the last inserted insert id.
        */
-      int lastInsertId(const std::string& name)
+      long lastInsertId(const std::string& name)
       { return conn->lastInsertId(name); }
 
       /**
